Auth(update) making the auth an molecule to use it in the Register
