--- front/src/pages/Auth/Login/styles.module.scss
@@ -1,34 +1,0 @@
-.container{
-    display: grid;
-    width: 100%;
-    height: 100%;
-
-    grid-template-rows: 100px 1fr;
-    grid-template-columns: 1fr;
-
-    .header{
-        grid-row: 1;
-        background-color: #00BBFF;
-        line-height: 100px;
-
-        &__title{
-            display: inline-block;
-            color: white;
-            font-size: 18px;
-            font-weight: bold;
-            vertical-align: middle;
-            padding-left: 30px;
-        }
-    }
-
-    .login{
-        grid-row: 2;
-        padding: 40px 0 0 30px;
-
-        form{
-            display: flex;
-            flex-direction: column;
-            gap: 15px;
-        }
-    }
-}
